# Speed up ActuateCollectorMetrics