# **Before the change**
1 thread 

```
Benchmark                                              Mode  Cnt    Score    Error  Units
MetricsBenchmarks.incrementBytes_longSpans_Actuate     avgt   15  295.329 ±  8.135  us/op
MetricsBenchmarks.incrementBytes_longSpans_inMemory    avgt   15    0.010 ±  0.001  us/op
MetricsBenchmarks.incrementBytes_mediumSpans_Actuate   avgt   15   60.035 ±  1.428  us/op
MetricsBenchmarks.incrementBytes_mediumSpans_inMemory  avgt   15    0.010 ±  0.001  us/op
MetricsBenchmarks.incrementBytes_shortSpans_Actuate    avgt   15   30.210 ±  1.112  us/op
MetricsBenchmarks.incrementBytes_shortSpans_inMemory   avgt   15    0.010 ±  0.001  us/op
```

# **After the change**

1 thread

```
Benchmark                                              Mode  Cnt  Score    Error  Units
MetricsBenchmarks.incrementBytes_longSpans_Actuate     avgt   15  0.108 ±  0.002  us/op
MetricsBenchmarks.incrementBytes_longSpans_inMemory    avgt   15  0.011 ±  0.001  us/op
MetricsBenchmarks.incrementBytes_mediumSpans_Actuate   avgt   15  0.108 ±  0.004  us/op
MetricsBenchmarks.incrementBytes_mediumSpans_inMemory  avgt   15  0.010 ±  0.001  us/op
MetricsBenchmarks.incrementBytes_shortSpans_Actuate    avgt   15  0.107 ±  0.002  us/op
MetricsBenchmarks.incrementBytes_shortSpans_inMemory   avgt   15  0.010 ±  0.001  us/op
```

4 threads
```
Benchmark                                              Mode  Cnt  Score    Error  Units
MetricsBenchmarks.incrementBytes_longSpans_Actuate     avgt   15  0.107 ±  0.012  us/op
MetricsBenchmarks.incrementBytes_longSpans_inMemory    avgt   15  0.014 ±  0.003  us/op
MetricsBenchmarks.incrementBytes_mediumSpans_Actuate   avgt   15  0.108 ±  0.003  us/op
MetricsBenchmarks.incrementBytes_mediumSpans_inMemory  avgt   15  0.011 ±  0.001  us/op
MetricsBenchmarks.incrementBytes_shortSpans_Actuate    avgt   15  0.107 ±  0.005  us/op
MetricsBenchmarks.incrementBytes_shortSpans_inMemory   avgt   15  0.012 ±  0.001  us/op
```

40 threads
```
Benchmark                                              Mode  Cnt  Score    Error  Units
MetricsBenchmarks.incrementBytes_longSpans_Actuate     avgt  240  0.755 ±  0.016  us/op
MetricsBenchmarks.incrementBytes_longSpans_inMemory    avgt  240  0.098 ±  0.001  us/op
MetricsBenchmarks.incrementBytes_mediumSpans_Actuate   avgt  240  0.786 ±  0.034  us/op
MetricsBenchmarks.incrementBytes_mediumSpans_inMemory  avgt  240  0.098 ±  0.001  us/op
MetricsBenchmarks.incrementBytes_shortSpans_Actuate    avgt  240  0.692 ±  0.006  us/op
MetricsBenchmarks.incrementBytes_shortSpans_inMemory   avgt  240  0.095 ±  0.001  us/op
```