@adriancole 2017-03-14
fantastic work! cc @dsyer @shakuzen 

this is per notes here https://github.com/openzipkin/zipkin/commit/568e3e3ffdd9f319c877692b2d268a894f95bef0#commitcomment-21300483

and this is in efforts to reduce the load on the zipkin server