@denyska 2017-03-14
Probably easiest way to test conditional logic is to enhance some of the Cassandra integration tests if any. I will take a look.