@adriancole 2017-03-14
yeah I'm mainly only concerned with a bean present and not (ex javaconfig
for counterbuffers/guagebuffers and another when they aren't there). you
can assert on the spring factory as a reality check (ex to verify
counterbuffers are or aren't there) before asking for collectormetrics