@adriancole 2017-03-20
released in 1.21.0